udpcast-20080914
v20080914
       * Added distclean target to make Debian build easier
       * Adapted to new name of mingw compiler, and other Mingw adaptations
       * Removed obsolete m486 flag
       * Fixed parameter types for getsockopt
       * If there are no participants after autostart delay, do not
       transmit but exit right away
       * RpmLint fixes
       * Adapt to Solaris 9 sparc
       * Added a lot of short opts (for systems without getopt_long)
       * pluggable rate governor
       * Mingw fixes (socket closing)

+++ fifo.c
@@ -5,7 +5,7 @@ void udpc_initFifo(struct fifo *fifo, int blockSize)
 {
     fifo->dataBufSize = blockSize * 4096;
     fifo->dataBuffer = xmalloc(fifo->dataBufSize+4096);
-    fifo->dataBuffer += 4096 - (((unsigned int)fifo->dataBuffer) % 4096);
+    fifo->dataBuffer += 4096 - (((unsigned long)fifo->dataBuffer) % 4096);
 
     /* Free memory queue is initially full */
     fifo->freeMemQueue = pc_makeProduconsum(fifo->dataBufSize, "free mem");
